replace old module
