+++ src/pykeen/metrics/classification.py
@@ -1,27 +1,82 @@
 # -*- coding: utf-8 -*-
 
-"""Classification metrics."""
+"""
+Classification metrics.
 
-import inspect
-from typing import Callable, ClassVar, MutableMapping, Optional, Type
+The metrics in this module assume the link prediction setting to be a (binary) classification of individual triples.
 
-import numpy as np
-import rexmex.metrics.classification as rmc
+.. note ::
+    many metrics in this module use `scikit-learn` under the hood, cf.
+    https://scikit-learn.org/stable/modules/model_evaluation.html#classification-metrics
+"""
+
+# see also: https://cran.r-project.org/web/packages/metrica/vignettes/available_metrics_classification.html
+
+from __future__ import annotations
+
+import abc
+import math
+import warnings
+from typing import ClassVar, Collection, Literal, Protocol
+
+import numpy
 from class_resolver import ClassResolver
+from sklearn import metrics
 
 from .utils import Metric, ValueRange
 
 __all__ = [
     "ClassificationMetric",
-    "classifier_annotator",
     "construct_indicator",
     "classification_metric_resolver",
 ]
 
-ClassificationFunc = Callable[[np.array, np.array], float]
+ZeroDivisionPolicy = Literal["warn", 0, 1]
 
 
-def construct_indicator(*, y_score: np.ndarray, y_true: np.ndarray) -> np.ndarray:
+def safe_divide(numerator: float, denominator: float, zero_division: ZeroDivisionPolicy = "warn") -> float:
+    """
+    Performs division and handles divide-by-zero similar to scikit-learn.
+
+    :param numerator:
+        the numerator
+    :param denominator:
+        the denominator
+    :param zero_division:
+        the zero-division policy; If "warn", act like 0, but warn about the case.
+    """
+    # todo: do we need numpy support?
+    if denominator != 0:
+        return numerator / denominator
+    if zero_division == "warn":
+        zero_division = 0
+        warnings.warn(
+            message=f"Division by zero. Result set to {zero_division} according to policy.", category=UserWarning
+        )
+    return zero_division
+
+
+class ClassificationFunc(Protocol):
+    """A protocol for classification functions."""
+
+    # TODO: check how to type properly
+    def __call__(self, y_true: numpy.ndarray, y_score: numpy.ndarray, /, **kwargs) -> float:
+        """
+        Calculate the metric.
+
+        :param y_true: shape: (num_samples,)
+            the true label, either 0 or 1.
+        :param y_score: shape: (num_samples,)
+            the predictions, either as continuous scores, or as binarized prediction
+            (depending on the concrete metric at hand).
+
+        :return:
+            a scalar metric value
+        """
+        ...
+
+
+def construct_indicator(*, y_score: numpy.ndarray, y_true: numpy.ndarray) -> numpy.ndarray:
     """Construct binary indicators from a list of scores.
 
     If there are $n$ positively labeled entries in ``y_true``, this function
@@ -48,153 +103,565 @@ def construct_indicator(*, y_score: np.ndarray, y_true: np.ndarray) -> np.ndarra
         This implementation was inspired by
         https://github.com/xptree/NetMF/blob/77286b826c4af149055237cef65e2a500e15631a/predict.py#L25-L33
     """
-    number_pos = np.sum(y_true, dtype=int)
-    y_sort = np.flip(np.argsort(y_score))
-    y_pred = np.zeros_like(y_true, dtype=int)
-    y_pred[y_sort[np.arange(number_pos)]] = 1
+    # TODO: re-consider threshold
+    number_pos = numpy.sum(y_true, dtype=int)
+    y_sort = numpy.flip(numpy.argsort(y_score))
+    y_pred = numpy.zeros_like(y_true, dtype=int)
+    y_pred[y_sort[numpy.arange(number_pos)]] = 1
     return y_pred
 
 
-class ClassificationMetric(Metric):
+class ClassificationMetric(Metric, abc.ABC):
     """A base class for classification metrics."""
 
-    #: A description of the metric
-    description: ClassVar[str]
     #: The function that runs the metric
     func: ClassVar[ClassificationFunc]
 
+    def __call__(self, y_true: numpy.ndarray, y_score: numpy.ndarray, weights: numpy.ndarray | None = None) -> float:
+        """Evaluate the metric.
+
+        :param y_true: shape: (num_samples,)
+            the true labels, either 0 or 1.
+        :param y_score: shape: (num_samples,)
+            the predictions, either continuous or binarized.
+        :param weights: shape: (num_samples,)
+            weights for individual predictions
+
+            .. warning ::
+                not all metrics support sample weights - check :attr:`supports_weights` first
+
+        :return:
+            the scalar metric value
+        """
+        if weights is None:
+            return self.func(y_true, y_score)
+        if not self.supports_weights:
+            raise ValueError(
+                f"{self.__call__.__qualname__} does not support sample weights but received" f"weights={weights}.",
+            )
+        return self.func(y_true, y_score, sample_weight=weights)
+
+
+def num_scores(y_true: numpy.ndarray, y_score: numpy.ndarray, /, **kwargs) -> float:
+    """Return the number of scores."""
+    return y_score.size
+
+
+class NumScores(ClassificationMetric):
+    """
+    The number of scores.
+
+    --
+    description: The number of scores.
+    """
+
+    func = num_scores
+
+
+class BinarizedClassificationMetric(ClassificationMetric, abc.ABC):
+    """A classification metric which requires binarized predictions instead of scores."""
+
+    binarize: ClassVar[bool] = True
+
     # docstr-coverage: inherited
-    @classmethod
-    def get_description(cls) -> str:  # noqa: D102
-        return cls.description
-
-    @classmethod
-    def score(cls, y_true, y_score) -> float:
-        """Run the scoring function."""
-        return cls.func(y_true, construct_indicator(y_score=y_score, y_true=y_true) if cls.binarize else y_score)
-
-
-#: Functions with the right signature in the :mod:`rexmex.metrics.classification` that are not themselves metrics
-EXCLUDE = {
-    rmc.true_positive,
-    rmc.true_negative,
-    rmc.false_positive,
-    rmc.false_negative,
-    rmc.pr_auc_score,  # for now there's an issue
-}
-
-#: This dictionary maps from duplicate functions to the canonical function in :mod:`rexmex.metrics.classification`
-DUPLICATE_CLASSIFIERS = {
-    rmc.miss_rate: rmc.false_negative_rate,
-    rmc.fall_out: rmc.false_positive_rate,
-    rmc.selectivity: rmc.true_negative_rate,
-    rmc.specificity: rmc.true_negative_rate,
-    rmc.hit_rate: rmc.true_positive_rate,
-    rmc.sensitivity: rmc.true_positive_rate,
-    rmc.critical_success_index: rmc.threat_score,
-    rmc.precision_score: rmc.positive_predictive_value,
-    rmc.recall_score: rmc.true_positive_rate,
-}
-
-
-class MetricAnnotator:
-    """A class for annotating metric functions."""
-
-    metrics: MutableMapping[str, Type[ClassificationMetric]]
-
-    def __init__(self):
-        """Initialize the annotator."""
-        self.metrics = {}
-
-    def higher(self, func: ClassificationFunc, **kwargs) -> None:
-        """Annotate a function where higher values are better."""
-        self.add(func, increasing=True, **kwargs)
-
-    def lower(self, func: ClassificationFunc, **kwargs) -> None:
-        """Annotate a function where lower values are better."""
-        self.add(func, increasing=False, **kwargs)
-
-    def add(
-        self,
-        func: ClassificationFunc,
-        *,
-        increasing: bool,
-        description: str,
-        link: str,
-        name: Optional[str] = None,
-        lower: Optional[float] = 0.0,
-        lower_inclusive: bool = True,
-        upper: Optional[float] = 1.0,
-        upper_inclusive: bool = True,
-        binarize: bool = False,
-    ) -> None:
-        """Annotate a function."""
-        title = func.__name__.replace("_", " ").title()
-        metric_cls = type(
-            title.replace(" ", ""),
-            (ClassificationMetric,),
-            dict(
-                name=name or title,
-                link=link,
-                binarize=binarize,
-                increasing=increasing,
-                value_range=ValueRange(
-                    lower=lower,
-                    lower_inclusive=lower_inclusive,
-                    upper=upper,
-                    upper_inclusive=upper_inclusive,
-                ),
-                func=func,
-                description=description,
-            ),
+    def __call__(
+        self, y_true: numpy.ndarray, y_score: numpy.ndarray, weights: numpy.ndarray | None = None
+    ) -> float:  # noqa: D102
+        return super().__call__(
+            y_true=y_true, y_score=construct_indicator(y_score=y_score, y_true=y_true), weights=weights
         )
-        self.metrics[func.__name__] = metric_cls
 
 
-classifier_annotator = MetricAnnotator()
+class BalancedAccuracyScore(BinarizedClassificationMetric):
+    """
+    The average of recall obtained on each class.
 
-classifier_annotator.higher(
-    rmc.balanced_accuracy_score,
-    binarize=True,
-    name="Balanced Accuracy",
-    description="An adjusted version of the accuracy for imbalanced datasets",
-    link="https://scikit-learn.org/stable/modules/generated/sklearn.metrics.balanced_accuracy_score.html",
-)
-classifier_annotator.higher(
-    rmc.matthews_correlation_coefficient,
-    binarize=True,
-    lower=-1.0,
-    upper=1.0,
-    description="A balanced measure applicable even with class imbalance",
-    link="https://en.wikipedia.org/wiki/Phi_coefficient",
-)
+    --
+    link: https://scikit-learn.org/stable/modules/generated/sklearn.metrics.balanced_accuracy_score.html
+    description: The average of recall obtained on each class.
+    """
+
+    value_range = ValueRange(lower=0, lower_inclusive=True, upper=1, upper_inclusive=True)
+    increasing: ClassVar[bool] = True
+    synonyms: ClassVar[Collection[str]] = ("b-acc", "bas")
+    supports_weights: ClassVar[bool] = True
+    func = metrics.balanced_accuracy_score
 
 
-classification_metric_resolver: ClassResolver[ClassificationMetric] = ClassResolver(
-    list(classifier_annotator.metrics.values()),
+class AveragePrecisionScore(ClassificationMetric):
+    """The average precision from prediction scores.
+
+    .. note ::
+        this metric is different from the area under the precision-recall curve, which uses
+        interpolation and can be too optimistic.
+
+    --
+    link: https://scikit-learn.org/stable/modules/generated/sklearn.metrics.average_precision_score.html
+    description: The average precision across different thresholds.
+    """
+
+    # TODO: can we directly include sklearn's docstring here?
+
+    name = "Average Precision Score (APS)"
+    value_range = ValueRange(lower=0, lower_inclusive=True, upper=1, upper_inclusive=True)
+    increasing: ClassVar[bool] = True
+    synonyms: ClassVar[Collection[str]] = ("aps", "ap")
+    supports_weights: ClassVar[bool] = True
+    func = metrics.average_precision_score
+
+
+class AreaUnderTheReceiverOperatingCharacteristicCurve(ClassificationMetric):
+    """
+    The area under the receiver operating characteristic curve.
+
+    --
+    link: https://scikit-learn.org/stable/modules/generated/sklearn.metrics.roc_auc_score.html
+    description: The area under the receiver operating characteristic curve.
+    """
+
+    name = "Area Under The Receiver Operating Characteristic Curve (ROC AUC)"
+    value_range = ValueRange(lower=0, lower_inclusive=True, upper=1, upper_inclusive=True)
+    increasing: ClassVar[bool] = True
+    synonyms: ClassVar[Collection[str]] = ("roc-auc",)
+    supports_weights: ClassVar[bool] = True
+    func = metrics.roc_auc_score
+
+
+class ConfusionMatrixClassificationMetric(ClassificationMetric, abc.ABC):
+    """A classification metric based on the confusion matrix."""
+
+    binarize: ClassVar[bool] = True
+    zero_division: ZeroDivisionPolicy = "warn"
+
+    @abc.abstractmethod
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:
+        """
+        Calculate the metric from the confusion table.
+
+        :param matrix: shape: (2, 2)
+            the confusion table
+                [[ TP, FN ]
+                 [ FP, TN ]]
+
+        :return:
+            the scalar metric
+        """
+        # todo: it would make sense to have a separate evaluator which constructs the confusion matrix only once
+        raise NotImplementedError
+
+    def __call__(self, y_true: numpy.ndarray, y_score: numpy.ndarray, weights: numpy.ndarray | None = None) -> float:
+        y_pred = construct_indicator(y_score=y_score, y_true=y_true)
+        matrix = metrics.confusion_matrix(y_true=y_true, y_pred=y_pred, sample_weight=weights, normalize=None)
+        return self.extract_from_confusion_matrix(matrix=matrix)
+
+
+class TruePositiveRate(ConfusionMatrixClassificationMetric):
+    """
+    The true positive rate is the probability that the prediction is positive, given the triple is truly positive.
+
+    .. math ::
+        TPR = TP / (TP + FN)
+
+    --
+    link: https://en.wikipedia.org/wiki/Sensitivity_(test)
+    description: The probability that a truly positive triple is predicted positive.
+    """
+
+    increasing: ClassVar[bool] = True
+    synonyms: ClassVar[Collection[str]] = ("tpr", "sensitivity", "recall", "hit rate")
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        return safe_divide(numerator=matrix[1, 1], denominator=matrix[1, :].sum(), zero_division=self.zero_division)
+
+
+class TrueNegativeRate(ConfusionMatrixClassificationMetric):
+    """
+    The true negative rate is the probability that the prediction is negative, given the triple is truly negative.
+
+    .. math ::
+        TNR = TN / (TN + FP)
+
+    .. warning ::
+        most knowledge graph datasets do not have true negatives, i.e., verified false facts, but rather are
+        collection of (mostly) true facts, where the missing ones are generally unknown rather than false.
+
+    --
+    link: https://en.wikipedia.org/wiki/Specificity_(tests)
+    description: The probability that a truly false triple is predicted negative.
+    """
+
+    increasing: ClassVar[bool] = True
+    synonyms: ClassVar[Collection[str]] = ("tnr", "specificity", "selectivity")
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        return safe_divide(numerator=matrix[0, 0], denominator=matrix[0, :].sum(), zero_division=self.zero_division)
+
+
+class FalsePositiveRate(ConfusionMatrixClassificationMetric):
+    """
+    The false positive rate is the probability that the prediction is positive, given the triple is truly negative.
+
+    .. math ::
+        FPR = FP / (FP + TN)
+
+    --
+    link: https://en.wikipedia.org/wiki/False_positive_rate
+    description: The probability that a truly negative triple is predicted positive.
+    """
+
+    increasing: ClassVar[bool] = False
+    synonyms: ClassVar[Collection[str]] = ("fpr", "fall-out", "false alarm ratio")
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        return safe_divide(numerator=matrix[1, 0], denominator=matrix[1, :].sum(), zero_division=self.zero_division)
+
+
+class FalseNegativeRate(ConfusionMatrixClassificationMetric):
+    """
+    The false negative rate is the probability that the prediction is negative, given the triple is truly positive.
+
+    .. math ::
+        FNR = FN / (FN + TP)
+
+    --
+    link: https://en.wikipedia.org/wiki/Type_I_and_type_II_errors#False_positive_and_false_negative_rates
+    description: The probability that a truly positive triple is predicted negative.
+    """
+
+    increasing: ClassVar[bool] = False
+    synonyms: ClassVar[Collection[str]] = ("fnr", "miss-rate")
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        return safe_divide(numerator=matrix[0, 1], denominator=matrix[0, :].sum(), zero_division=self.zero_division)
+
+
+class PositivePredictiveValue(ConfusionMatrixClassificationMetric):
+    """
+    The positive predictive value is the proportion of predicted positives which are true positive.
+
+    .. math ::
+        PPV = TP / (TP + FP)
+
+    --
+    link: https://en.wikipedia.org/wiki/Positive_and_negative_predictive_values
+    description: The proportion of predicted positives which are true positive.
+    """
+
+    increasing: ClassVar[bool] = True
+    synonyms: ClassVar[Collection[str]] = ("ppv",)
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        return safe_divide(numerator=matrix[0, 0], denominator=matrix[:, 0].sum(), zero_division=self.zero_division)
+
+
+class NegativePredictiveValue(ConfusionMatrixClassificationMetric):
+    """
+    The negative predictive value is the proportion of predicted negatives which are true negative.
+
+    .. math ::
+        NPV = TN / (TN + FN)
+
+    --
+    link: https://en.wikipedia.org/wiki/Positive_and_negative_predictive_values
+    description: The proportion of predicted negatives which are true negatives.
+    """
+
+    increasing: ClassVar[bool] = True
+    synonyms: ClassVar[Collection[str]] = ("npv",)
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        return safe_divide(numerator=matrix[1, 1], denominator=matrix[:, 1].sum(), zero_division=self.zero_division)
+
+
+class FalseDiscoveryRate(ConfusionMatrixClassificationMetric):
+    """
+    The false discovery rate is the proportion of predicted negatives which are true positive.
+
+    .. math ::
+        FDR = FP / (FP + TP)
+
+    --
+    link: https://en.wikipedia.org/wiki/False_discovery_rate
+    description: The proportion of predicted negatives which are true positive.
+    """
+
+    increasing: ClassVar[bool] = False
+    synonyms: ClassVar[Collection[str]] = ("fdr",)
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        return safe_divide(numerator=matrix[1, 0], denominator=matrix[:, 0].sum(), zero_division=self.zero_division)
+
+
+class FalseOmissionRate(ConfusionMatrixClassificationMetric):
+    """
+    The false omission rate is the proportion of predicted positives which are true negative.
+
+    .. math ::
+        FOR = FN / (FN + TN)
+
+    --
+    link: https://en.wikipedia.org/wiki/False_discovery_rate
+    description: The proportion of predicted positives which are true negative.
+    """
+
+    increasing: ClassVar[bool] = False
+    synonyms: ClassVar[Collection[str]] = ("fom",)
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        return safe_divide(numerator=matrix[0, 1], denominator=matrix[:, 1].sum(), zero_division=self.zero_division)
+
+
+class PositiveLikelihoodRatio(ConfusionMatrixClassificationMetric):
+    r"""
+    The positive likelihood ratio is the ratio of true positive rate to false positive rate.
+
+    .. math ::
+        LR+ = TPR / FPR = \frac{TP / (TP + FN)}{FP / (FP + TN)} = \frac{TP \cdot (FP + TN)}{FP \cdot (TP + FN)}
+
+    --
+    link: https://en.wikipedia.org/wiki/Likelihood_ratios_in_diagnostic_testing#Calculation_example
+    description: The ratio of true positive rate to false positive rate.
+    """
+
+    increasing: ClassVar[bool] = True
+    synonyms: ClassVar[Collection[str]] = ("lr+",)
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        return safe_divide(
+            numerator=matrix[0, 0] * matrix[1, :].sum(),
+            denominator=matrix[1, 0] * matrix[0, :].sum(),
+            zero_division=self.zero_division,
+        )
+
+
+class NegativeLikelihoodRatio(ConfusionMatrixClassificationMetric):
+    r"""
+    The negative likelihood ratio is the ratio of false negative rate to true negative rate.
+
+    .. math ::
+        LR- = FNR / TNR = \frac{FN / (TP + FN)}{TN / (FP + TN)} = \frac{FN \cdot (FP + TN)}{TN \cdot (TP + FN)}
+
+    --
+    link: https://en.wikipedia.org/wiki/Likelihood_ratios_in_diagnostic_testing#Calculation_example
+    description: The ratio of false positive rate to true positive rate.
+    """
+
+    increasing: ClassVar[bool] = False
+    synonyms: ClassVar[Collection[str]] = ("lr-",)
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        return safe_divide(
+            numerator=matrix[0, 1] * matrix[0, :].sum(),
+            denominator=matrix[1, 1] * matrix[1, :].sum(),
+            zero_division=self.zero_division,
+        )
+
+
+class DiagnosticOddsRatio(ConfusionMatrixClassificationMetric):
+    r"""
+    The ratio of positive and negative likelihood ratio.
+
+    .. math ::
+        DOR = \frac{LR+}{LR-} = \frac{TP \cdot TN}{FP \cdot FN}
+
+    --
+    link: https://en.wikipedia.org/wiki/Diagnostic_odds_ratio
+    description: The ratio of positive and negative likelihood ratio.
+    """
+
+    increasing: ClassVar[bool] = True
+    synonyms: ClassVar[Collection[str]] = ("lr-",)
+
+    # todo: https://en.wikipedia.org/wiki/Diagnostic_odds_ratio#Confidence_interval
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        return safe_divide(
+            numerator=matrix[0, 0] * matrix[1, 1],
+            denominator=matrix[0, 1] * matrix[1, 0],
+            zero_division=self.zero_division,
+        )
+
+
+class Accuracy(ConfusionMatrixClassificationMetric):
+    r"""
+    The ratio of the number of correct classifications to the total number.
+
+    .. math ::
+        ACC = (TP + TN) / (TP + TN + FP + FN)
+
+    --
+    link: https://en.wikipedia.org/wiki/Evaluation_of_binary_classifiers#Single_metrics
+    description: The ratio of the number of correct classifications to the total number.
+    """
+    increasing: ClassVar[bool] = True
+    synonyms: ClassVar[Collection[str]] = ("acc", "fraction correct", "fc")
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        return safe_divide(
+            numerator=matrix[0, 0] * matrix[1, 1], denominator=matrix.sum(), zero_division=self.zero_division
+        )
+
+
+class F1Score(ConfusionMatrixClassificationMetric):
+    r"""
+    The harmonic mean of precision and recall.
+
+    .. math ::
+        F1 = 2TP / (2TP + FP + FN)
+
+    --
+    link: https://en.wikipedia.org/wiki/F1_score
+    description: The harmonic mean of precision and recall.
+    """
+    increasing: ClassVar[bool] = True
+    synonyms: ClassVar[Collection[str]] = ("f1",)
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        return safe_divide(
+            numerator=2 * matrix[0, 0],
+            denominator=2 * matrix[0, 0] + matrix[0, 1] + matrix[1, 0],
+            zero_division=self.zero_division,
+        )
+
+
+class PrevalenceThreshold(ConfusionMatrixClassificationMetric):
+    r"""
+    The prevalence threshold.
+
+    .. math ::
+        PT = √FPR / (√TPR + √FPR)
+
+    --
+    link: https://en.wikipedia.org/wiki/Prevalence_threshold
+    description: The prevalence threshold.
+    """
+    # todo: improve doc
+    increasing: ClassVar[bool] = False
+    synonyms: ClassVar[Collection[str]] = ("pt",)
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        fpr = FalsePositiveRate().extract_from_confusion_matrix(matrix=matrix)
+        tpr = TruePositiveRate().extract_from_confusion_matrix(matrix=matrix)
+        return safe_divide(
+            numerator=numpy.sqrt(fpr),
+            denominator=numpy.sqrt(fpr) + numpy.sqrt(tpr),
+            zero_division=self.zero_division,
+        )
+
+
+class ThreatScore(ConfusionMatrixClassificationMetric):
+    r"""
+    The threat score.
+
+    .. math ::
+        TS = TP / (TP + FN + FP)
+
+    --
+    link: https://en.wikipedia.org/wiki/Prevalence_threshold
+    description: The harmonic mean of precision and recall.
+    """
+    increasing: ClassVar[bool] = True
+    synonyms: ClassVar[Collection[str]] = ("ts", "critical success index", "csi", "jaccard index")
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        return safe_divide(
+            numerator=matrix[0, 0],
+            denominator=matrix[0, 0] + matrix[0, 1] + matrix[1, 0],
+            zero_division=self.zero_division,
+        )
+
+
+class FowlkesMallowsIndex(ConfusionMatrixClassificationMetric):
+    r"""
+    The Fowlkes Mallows index.
+
+    .. math ::
+        FM = \sqrt{\frac{TP^2}{(2TP + FP + FN)}}
+
+    --
+    link: https://en.wikipedia.org/wiki/Prevalence_threshold
+    description: The Fowlkes Mallows index.
+    """
+    increasing: ClassVar[bool] = True
+    synonyms: ClassVar[Collection[str]] = ("fm", "fmi")
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        return math.sqrt(
+            safe_divide(
+                numerator=matrix[0, 0] ** 2,
+                denominator=2 * matrix[0, 0] + matrix[0, 1] + matrix[1, 0],
+                zero_division=self.zero_division,
+            )
+        )
+
+
+class Informedness(ConfusionMatrixClassificationMetric):
+    r"""
+    The informedness metric.
+
+    .. math ::
+        YI = TPR + TNR - 1 = TP / (TP + FN) + TN / (TN + FP) - 1
+
+    --
+    link: https://en.wikipedia.org/wiki/Informedness
+    description: The informedness metric.
+    """
+    increasing: ClassVar[bool] = True
+    synonyms: ClassVar[Collection[str]] = ("Youden's J", "Youden's Index", "yi")
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        return (
+            safe_divide(numerator=matrix[1, 1], denominator=matrix[1, :].sum(), zero_division=self.zero_division)
+            + safe_divide(numerator=matrix[0, 0], denominator=matrix[0, :].sum(), zero_division=self.zero_division)
+            - 1
+        )
+
+
+class MatthewsCorrelationCoefficient(ConfusionMatrixClassificationMetric):
+    r"""
+    The Matthews Correlation Coefficient (MCC).
+
+    A balanced measure applicable even with class imbalance.
+
+    .. math ::
+        MCC = (TP * TN - FP * FN) / sqrt((TP + FP) * (TP + FN) * (TN + FP) * (TN + FN))
+
+    --
+    link: https://en.wikipedia.org/wiki/Phi_coefficient
+    description: The Matthews Correlation Coefficient (MCC).
+    """
+    increasing: ClassVar[bool] = True
+    synonyms: ClassVar[Collection[str]] = ("mcc",)
+    value_range: ClassVar[ValueRange] = ValueRange(lower=-1, upper=1, lower_inclusive=True, upper_inclusive=True)
+
+    # docstr-coverage: inherited
+    def extract_from_confusion_matrix(self, matrix: numpy.ndarray) -> float:  # noqa: D102
+        return safe_divide(
+            numerator=matrix[0, 0] * matrix[1, 1] - matrix[1, 0] * matrix[0, 1],
+            denominator=matrix.sum(axis=1).prod() * matrix.sum(axis=0).prod(),
+            zero_division=self.zero_division,
+        )
+
+
+classification_metric_resolver: ClassResolver[ClassificationMetric] = ClassResolver.from_subclasses(
     base=ClassificationMetric,
+    default=AveragePrecisionScore,
+    skip={BinarizedClassificationMetric, ConfusionMatrixClassificationMetric},
 )
-
-
-def _check():
-    """Check that all functions in the classification module are annotated."""
-    for func in rmc.__dict__.values():
-        if not inspect.isfunction(func):
-            continue
-        parameters = inspect.signature(func).parameters
-        if "y_true" not in parameters or "y_score" not in parameters:
-            continue
-        if func in EXCLUDE:
-            if func in classifier_annotator.metrics:
-                raise ValueError(f"should not include {func.__name__}")
-            continue
-        if func in DUPLICATE_CLASSIFIERS:
-            if func in classifier_annotator.metrics:
-                raise ValueError(f"{func.__name__} is a duplicate of {DUPLICATE_CLASSIFIERS[func].__name__}")
-            continue
-        if func.__name__ not in classifier_annotator.metrics:
-            raise ValueError(f"missing rexmex classifier: {func.__name__}")
-
-
-if __name__ == "__main__":
-    _check()
